Server Update
Did some quality of life updates

+++ Multiplayer Server/src/GUI/Connection.java
@@ -8,11 +8,11 @@ import javax.swing.JButton;
 public class Connection {
 	public final int ID;
 	public final String ADDRESS;
-	public int REFRESH_RATE;
+	public long REFRESH_RATE;
 	public String CHARACTER_CLASS, USERNAME;
 	public final JButton BUTTON;
 
-	public Connection(int id, String address, String username, int refreshRate, String characterClass) {
+	public Connection(int id, String address, String username, long refreshRate, String characterClass) {
 		ID = id;
 		ADDRESS = address;
 		USERNAME = username;
